商品接口调整

+++ yshop-pms/pms-service/src/main/java/com/fly4j/yshop/pms/controller/app/AppGoodsController.java
@@ -47,7 +47,7 @@ public class AppGoodsController {
     })
     @GetMapping()
     public R list(
-            @RequestParam(required = false) Integer category_id,
+            @RequestParam(required = false) Long category_id,
             @RequestParam(required = false) Integer limit,
             @RequestParam(required = false) Long is_hot,
             @RequestParam(required = false) Long is_new
